Refactor StatisticCard Component to Common Directory with Comments and Documentation
Related to #1462
