+++ src/app/components/common/index.ts
@@ -1,0 +1,5 @@
+import { StatisticCard } from "./StatisticCard";
+
+export {
+  StatisticCard,
+}
