+++ src/app/components/dashboard/index.ts
@@ -1,11 +1,9 @@
 import { NavBar } from "./NavBar";
 import { BarChart } from "./BarChart";
-import { StatisticCard } from "./StatisticCard";
 import { AffiliatesList } from "./AffiliatesList";
 
 export {
   NavBar,
   BarChart,
-  StatisticCard,
   AffiliatesList
 }
